Switch back to generating python protocol buffer files with Makefile
Android studio complains about the ../../../../../../server path that
points outside of the project tree. Aside from that we shouldn't have to
install a complete java/gradle/android-studio environment to build the
Python OpenRTiST server or client. Also, we already need the Makefile
and python virtualenv anyway to freeze the models we need for on-device
execution.